fix type mismatch

+++ CoordgenMacrocycleBuilder.cpp
@@ -719,9 +719,9 @@ sketcherMinimizerBond*
 CoordgenMacrocycleBuilder::findBondToOpen(sketcherMinimizerRing* ring) const
 {
     sketcherMinimizerBond* bestBond = nullptr;
-    size_t bestScore = 0;
+    float bestScore = 0.f;
     for (sketcherMinimizerBond* bond : ring->_bonds) {
-        size_t score = 0;
+        float score = 0.f;
         if (ring->isMacrocycle()) {
             if (bond->getBondOrder() != 1) {
                 continue;
